apps/membergpt-api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
openai
pdfplumber
python-multipart
requests